Refactor code and fixes #108
